+++ packages/ra-core/src/fetch/useMutation.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 
-import { useSafeSetState } from './hooks';
+import { useSafeSetState } from '../util/hooks';
 import useDataProvider from './useDataProvider';
 
 export interface Query {
